server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -3,12 +3,13 @@ const { Pool } = require("pg");
 const InvariantError = require("../../exceptions/InvariantError");
 const { mapDBToModel } = require("../../utils");
 const NotFoundError = require("../../exceptions/NotFoundError");
-const AuthorizationError = require('../../exceptions/AuthorizationError');
+const AuthorizationError = require("../../exceptions/AuthorizationError");
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   async addNote({ title, body, tags, owner }) {
@@ -27,23 +28,30 @@ class NotesService {
       throw new InvariantError("Catatan gagal ditambahkan");
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
+
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: `SELECT notes.* FROM notes
-            LEFT JOIN collaborations ON collaborations.note_id = notes.id
-            WHERE notes.owner = $1 OR collaborations.user_id = $1
-            GROUP BY notes.id`,
-      values: [owner],
-    };
+    try {
+      const result = await this._cacheService.get(`notes:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      const query = {
+        text: `SELECT notes.* FROM notes
+              LEFT JOIN collaborations ON collaborations.note_id = notes.id
+              WHERE notes.owner = $1 OR collaborations.user_id = $1
+              GROUP BY notes.id`,
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    if (!result.rowCount) {
-      return owner;
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      await this._cacheService.set(`notes:${owner}`, JSON.stringify(mappedResult));
+      return mappedResult;
     }
-    return result.rows.map(mapDBToModel);
   }
 
   async getNoteById(id) {
@@ -57,7 +65,7 @@ class NotesService {
     const result = await this._pool.query(query);
 
     if (!result.rows.length) {
-      throw new NotFoundError('Catatan tidak ditemukan');
+      throw new NotFoundError("Catatan tidak ditemukan");
     }
 
     return result.rows.map(mapDBToModel)[0];
@@ -66,46 +74,52 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id",
+      text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner",
       values: [title, body, tags, updatedAt, id],
     };
 
     const result = await this._pool.query(query);
 
     if (!result.rows.length) {
-      throw new NotFoundError('Gagal memperbarui catatan. Id tidak ditemukan');
+      throw new NotFoundError("Gagal memperbarui catatan. Id tidak ditemukan");
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+      text: "DELETE FROM notes WHERE id = $1 RETURNING id, owner",
       values: [id],
     };
- 
+
     const result = await this._pool.query(query);
- 
+
     if (!result.rows.length) {
-      throw new NotFoundError('Catatan gagal dihapus. Id tidak ditemukan');
+      throw new NotFoundError("Catatan gagal dihapus. Id tidak ditemukan");
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
     const query = {
-      text: 'SELECT * FROM notes WHERE id = $1',
+      text: "SELECT * FROM notes WHERE id = $1",
       values: [id],
     };
     const result = await this._pool.query(query);
     if (!result.rows.length) {
-      throw new NotFoundError('Catatan tidak ditemukan');
+      throw new NotFoundError("Catatan tidak ditemukan");
     }
     const note = result.rows[0];
     if (note.owner !== owner) {
-      throw new AuthorizationError('Anda tidak berhak mengakses resource ini');
+      throw new AuthorizationError("Anda tidak berhak mengakses resource ini");
     }
   }
 
-  async verifyNoteAccess(noteId, userId){
+  async verifyNoteAccess(noteId, userId) {
     try {
       await this.verifyNoteOwner(noteId, userId);
     } catch (error) {
